requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
scikit-learn
xgboost
joblib
httpx
tensorflow-cpu
